+++ lib/unisec.rb
@@ -2,9 +2,10 @@
 
 require 'unisec/version'
 
-require 'unisec/surrogates'
+require 'unisec/confusables'
 require 'unisec/hexdump'
 require 'unisec/properties'
-require 'unisec/confusables'
-require 'unisec/versions'
+require 'unisec/rugrep'
 require 'unisec/size'
+require 'unisec/surrogates'
+require 'unisec/versions'